Apply changes to support move
