+++ test/controllers/admin/health/airbrake.controller.test.js
@@ -7,7 +7,7 @@ const { describe, it, before, beforeEach, after } = exports.lab = Lab.script()
 const { expect } = Code
 
 // For running our service
-const { deployment } = require('../../server')
+const { deployment } = require('../../../../server')
 
 // Test helpers
 const { AuthorisationHelper, DatabaseHelper } = require('../support/helpers')
@@ -39,7 +39,7 @@ describe('Airbrake controller: GET /status/airbrake', () => {
   it('returns a 500 error', async () => {
     const options = {
       method: 'GET',
-      url: '/status/airbrake',
+      url: '/admin/health/airbrake',
       headers: { authorization: `Bearer ${authToken}` }
     }
 
@@ -50,7 +50,7 @@ describe('Airbrake controller: GET /status/airbrake', () => {
   it('causes Airbrake to send a notification', async () => {
     const options = {
       method: 'GET',
-      url: '/status/airbrake',
+      url: '/admin/health/airbrake',
       headers: { authorization: `Bearer ${authToken}` }
     }
